Do not leak InaccessibleObjectException from StringFastPath
Java 9 added a new exception which can be thrown by
Field.setAccessible(true). Unfortunately, that new exception inherits
directly from RuntimeException, instead of ReflectiveOperationException,
and it does not exist on Java 8, making it harder to catch and ignore.

However, Java 9 also added a replacement to Field.setAccessible(true)
which does not throw that new exception; instead of throwing that
exception, it does nothing, which will make the following operation
throw an IllegalAccessException, which is already correctly ignored.

Use that replacement when available, to make StringFastPath more robust
on Java 9 and newer. It still does not work correctly (it will treat all
String as non-compact), but at least will not break users of this class.